[Feat/#124] change the homeviewState after you apply LTCI
homeviewmodel을 만들고, SubmitView가 homeviewmodel를 의존한다.
단, SubmitView는 SubmitCheckListView를 통해 생성되기 때문에
SubmitListView도 의존성을 가져야 된다는 문제가 있다
