+++ Onbom/Onbom/View/LTCI/MainView.swift
@@ -18,29 +18,6 @@ struct MainView: View {
             case .home:
                 NavigationStack(path: $homeNavigation.homePath) {
                     HomeView()
-                        .navigationDestination(for: HomeRoute.self) { route in
-                            switch(route) {
-                            case .OnboardingView:                   OnboardingView(isOnboarding: .constant(true))
-                            case .ApplyHistoryView:                 ApplyHistoryView()
-                            case .DescriptionView:                  DescriptionView()
-                            case .ApplyTypeView:                    ApplyTypeView().toolbar(.hidden, for: .tabBar)
-                            case .MediHistoryView:                  MediHistoryView()
-                            case .MediConditionView:                MediConditionView()
-                            case .IDCardDescriptionView:            IDCardDescriptionView()
-                            case .IDCardConfirmEditView:            IDCardConfirmEditView()
-                            case .AddressFormView_Patient:          AddressFormView(formType: .patient)
-                            case .AddressFormView_ActualPatient:    AddressFormView(formType: .actualPatient)
-                            case .AddressFormView_Agent:            AddressFormView(formType: .agent)
-                            case .SignatureView:                    SignatureView()
-                            case .SubmitCheckListView:              SubmitCheckListView()
-                            case .StepView_First:                   StepView(state: .FIRST)
-                            case .StepView_Second:                  StepView(state: .SECOND)
-                            case .PatientInfoView:                  PatientInfoView()
-                            case .AgentInfoView:                    AgentInfoView()
-                            case .RejectView:                       RejectView()
-                            default:                                RejectView()
-                            }
-                        }
                 }
                 .environmentObject(homeNavigation)
                 .environmentObject(pdfManager)
