+++ Onbom/Onbom/View/LTCI/SubmitCheckListView.swift
@@ -11,6 +11,7 @@ struct SubmitCheckListView: View {
     @EnvironmentObject var patient: Patient
     @EnvironmentObject var agent: Agent
     @EnvironmentObject var pdfManager: PDFManager
+    @ObservedObject var homeViewModel: HomeViewModel
     @State private var isSubmitLoadingViewPresented = false
     
     var body: some View {
@@ -249,7 +250,7 @@ struct SubmitCheckListView: View {
         .padding([.top, .leading, .trailing], 20)
         .navigationBarBackButton()
         .fullScreenCover(isPresented: $isSubmitLoadingViewPresented) {
-            SubmitView()
+            SubmitView(homeViewModel: self.homeViewModel)
             
         }
     }
@@ -257,7 +258,7 @@ struct SubmitCheckListView: View {
 
 struct SubmitCheckListView_Previews: PreviewProvider {
     static var previews: some View {
-        SubmitCheckListView()
+        SubmitCheckListView(homeViewModel: HomeViewModel())
             .environmentObject(Patient())
             .environmentObject(Agent())
     }
